reduce differences in remote join, deprecate Translate module
