+++ server/src-lib/Hasura/GraphQL/Execute/LiveQuery/Plan.hs
@@ -26,28 +26,28 @@ module Hasura.GraphQL.Execute.LiveQuery.Plan
 
 import           Hasura.Prelude
 
-import qualified Data.Aeson.Casing                             as J
-import qualified Data.Aeson.Extended                           as J
-import qualified Data.Aeson.TH                                 as J
-import qualified Data.ByteString                               as B
-import qualified Data.HashMap.Strict                           as Map
-import qualified Data.HashMap.Strict.InsOrd                    as OMap
-import qualified Data.Sequence                                 as Seq
-import qualified Data.Text                                     as T
-import qualified Data.UUID.V4                                  as UUID
-import qualified Database.PG.Query                             as Q
-import qualified Database.PG.Query.PTI                         as PTI
-import qualified Language.GraphQL.Draft.Syntax                 as G
-import qualified PostgreSQL.Binary.Encoding                    as PE
+import qualified Data.Aeson.Casing                           as J
+import qualified Data.Aeson.Extended                         as J
+import qualified Data.Aeson.TH                               as J
+import qualified Data.ByteString                             as B
+import qualified Data.HashMap.Strict                         as Map
+import qualified Data.HashMap.Strict.InsOrd                  as OMap
+import qualified Data.Sequence                               as Seq
+import qualified Data.Text                                   as T
+import qualified Data.UUID.V4                                as UUID
+import qualified Database.PG.Query                           as Q
+import qualified Database.PG.Query.PTI                       as PTI
+import qualified Language.GraphQL.Draft.Syntax               as G
+import qualified PostgreSQL.Binary.Encoding                  as PE
 
 import           Control.Lens
-import           Data.UUID                                     (UUID)
+import           Data.UUID                                   (UUID)
 
-import qualified Hasura.Backends.Postgres.SQL.DML              as S
-import qualified Hasura.Backends.Postgres.Translate.RemoteJoin as RR
-import qualified Hasura.Backends.Postgres.Translate.Select     as DS
-import qualified Hasura.GraphQL.Parser.Schema                  as PS
-import qualified Hasura.RQL.IR.Select                          as DS
+import qualified Hasura.Backends.Postgres.Execute.RemoteJoin as RR
+import qualified Hasura.Backends.Postgres.SQL.DML            as S
+import qualified Hasura.Backends.Postgres.Translate.Select   as DS
+import qualified Hasura.GraphQL.Parser.Schema                as PS
+import qualified Hasura.RQL.IR.Select                        as DS
 
 import           Hasura.Backends.Postgres.Connection
 import           Hasura.Backends.Postgres.SQL.Error
